+++ src/main/kotlin/io/boxtime/cli/adapters/h2taskdatabase/TaskEntity.kt
@@ -8,6 +8,7 @@ import org.springframework.data.domain.Persistable
 import org.springframework.data.relational.core.mapping.Column
 import org.springframework.data.relational.core.mapping.MappedCollection
 import org.springframework.data.relational.core.mapping.Table
+import java.time.LocalDateTime
 
 @Table("TASK")
 data class TaskEntity(
@@ -16,6 +17,8 @@ data class TaskEntity(
     private val id: String,
     @Column("TITLE")
     val title: String,
+    @Column("CREATED_DATE")
+    val created: LocalDateTime,
     @MappedCollection(idColumn = "TASK_ID")
     val tags: Set<TagRef> = HashSet(),
     @Transient
@@ -23,14 +26,14 @@ data class TaskEntity(
 ) : Persistable<String> {
 
     @PersistenceCreator
-    constructor(id: String, title: String, tags: Set<TagRef>) : this(id, title, tags, false)
+    constructor(id: String, title: String, created: LocalDateTime, tags: Set<TagRef>) : this(id, title, created, tags, false)
 
     companion object {
         fun fromDomainObject(task: Task, new: Boolean): TaskEntity {
             val tagRefs = task.tags
                 .map { TagRef(it.id) }
                 .toSet()
-            return TaskEntity(task.id, task.title, tagRefs, new)
+            return TaskEntity(task.id, task.title, task.created, tagRefs, new)
         }
     }
 
@@ -38,7 +41,7 @@ data class TaskEntity(
         val tags = tagRepository.findByIdIn(this.tags.map { it.tagId })
             .map { it.toDomainObject() }
             .toSet()
-        return Task(this.id, this.title, tags)
+        return Task(this.id, this.title, this.created, tags)
     }
 
     override fun getId(): String {